[chore] cleanup and feedback
