+++ bin/report_python_modules
@@ -4,6 +4,9 @@ import json
 from sys import argv
 from os import chdir
 
+# When run with no arguments, enumerate the dependencies in
+# ./requirements.txt; given an argument, enumerate the dependencies listed in
+# the requirements.txt in that directory
 if len(argv) > 1:
     chdir(argv[1])
