+++ lib/salus.rb
@@ -17,7 +17,7 @@ require 'salus/config'
 require 'salus/processor'
 
 module Salus
-  VERSION = '1.0.0'.freeze
+  VERSION = '2.0.0'.freeze
   DEFAULT_REPO_PATH = './repo'.freeze # This is inside the docker container at /home/repo.
 
   SafeYAML::OPTIONS[:default_mode] = :safe
