+++ lib/salus/scanners/pattern_search.rb
@@ -93,7 +93,7 @@ module Salus::Scanners
 
       report_info(:hits, all_hits)
       errors.each { |error| report_error('Call to sift failed', error) }
-      failure_messages.each { |message| report_error(message) }
+      report_info(:failure_messages, failure_messages)
 
       if errors.empty? && failure_messages.empty?
         report_success
