+++ spec/fixtures/docker/expected_report.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.0.0",
+  "version": "2.0.0",
   "passed": true,
   "scans": {
     "PatternSearch": {
@@ -15,6 +15,9 @@
             "msg": "",
             "hit": "README.md:1:# placeholder"
           }
+        ],
+        "failure_messages": [
+
         ]
       },
       "errors": [
