+++ spec/fixtures/processor/local_uri/expected_report.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.0.0",
+  "version": "2.0.0",
   "passed": true,
   "scans": {
     "PatternSearch": {
@@ -8,6 +8,9 @@
       "running_time": 0.0,
       "info": {
         "hits": [
+
+        ],
+        "failure_messages": [
 
         ]
       },
